+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## v3.4.1
 
-- Claude: adapt selector for 'recent chats' element in sidebar
+- Claude: adapt selector for 'recent chats' element in sidebar or focus the first button or link
 
 ### On Progress